sessions: move IAgentHostFilterService out of providers layer
Fixes #316233
